app/requirements.txt
Flask==2.2.2
psycopg2-binary